Hint

Description

utility functions related to the Haskell Interpreter hint

Documentation

runInterpreterMain :: Interpreter () -> IO ()

run an interpreter monad, printing errors, ignoring values

runInterpreter_ :: Interpreter a -> IO (Either InterpreterError a)

run an interpreter monad with imports

say :: String -> Interpreter ()

print in the Interpreter monad

errorString :: InterpreterError -> String

run-time Language.Haskell.Interpreter compilation error

showError :: GhcError -> String

show a GHC error

interpretIO :: String -> Interpreter String

interpret a stringified IO command

fnIoPairs :: Int -> Expr -> Expr -> Interpreter String

get input-output pairs for a function given the inputs (for one concrete input type instantiation). | function application is run trough try-evaluate so as to Either-wrap potential run-time errors for partial functions. | the reason this function needs to be run through the interpreter is I only have the functioninputs as ASTstrings.

genInputs :: Int -> Tp -> Interpreter [Expr]

generate examples given a concrete type | the reason this function needs to be run through the interpreter is Gen wants to know its type at compile-time.

exprType :: Expr -> Interpreter Tp

get the type of an expression