Orphanage

Description

defined type class instances

Orphan instances

Hashable (Type l)
ensure I can use types as keys in HashMaps

Instance details

Methods

hashWithSalt :: Int -> Type l -> Int

hash :: Type l -> Int

Hashable (Exp l)
ensure I can use expressions as keys in HashMaps

Instance details

Methods

hashWithSalt :: Int -> Exp l -> Int

hash :: Exp l -> Int