Synthesis.Ast

Description

ast manipulation

Documentation

skeleton :: Tp -> Expr

make a typed hole for a type

hasHoles :: Expr -> Bool

check if an expression contains holes

holeExpr :: Expr

replace any holes in an expression with undefined, for type-checking purposes

hole `_` as an AST Expr

numAstNodes :: Expr -> Int

number of AST nodes in an Expr

letRes :: Expr -> Expr

get the body of a let-in expression

genBlockVariants :: HashMap String Tp -> [(String, Expr)]

generate applied variants of a function, e.g. [id, `id _`]

anyFn :: Expr

_ :: (_ -> _)

genUncurry :: Int -> Expr

generate an expression for an n-ary uncurry function, e.g. for n=2: ` fn (a, b) -> fn a b`

genFnType :: IO Tp

generate a function type, to then generate functions matching this type | deprecated, not in use

genFnInType :: IO Tp

generate a parameter type, to then generate functions taking this input | deprecated, not in use

genInputs :: Int -> Tp -> [Expr]

randomly generate a number of samples for a given type. | this should cover types from typesByArity.