Synthesis.Configs

Documentation

seed :: Int

random seed | TODO: make this program input

typesByArity :: HashMap Int [String]

synthesized types, categorized by arity

crashOnError :: Bool

when calculating function outputs either (False) performing an additional typecheck (slower), or (True) just crashing on error

nestLimit :: Int

max number of levels of nesting for generated types. | high values make for big logs while debugging...

maxInstances :: Int

max number of instantiations to generate for any type containing type variables. | may get less after nub filters out duplicate type instances.

numInputs :: Int

max number of inputs to generate. | may get less after nub filters out duplicates.

maxWildcardDepth :: Int

the maximum level of functions to imagine in a wildcard for function generation

genMaxHoles :: Int

the maximum number of holes to allow in a generated expression

split :: (Double, Double, Double)

ratio between the number of samples to use in the training, validation, and test sets, respectively.

listLengths :: (Int, Int)

the range of elements to generate for list types

intRange :: (Integer, Integer)

the range of integers to generate