Synthesis.FindHoles

Description

find holes in an AST

Documentation

gtrExpr :: Exp l -> Exp l

get the first sub-expression

strExpr :: Exp l -> Exp l -> Exp l

set the first sub-expression

findHolesExpr :: Exp l1 -> [(Exp l2 -> Exp l2, Exp l3 -> Exp l3 -> Exp l3)]

look for holes in an expression. to simplify extracting type, we will only look for holes as part of an ExpTypeSig, e.g. `_ :: Bool` | I couldn't figure out how to get this to typecheck as a whole lens, so instead I'm taking them as getter/setter pairs...