Synthesis.Generation

Description

generate task functions and sample input/output pairs

Documentation

fnOutputs

Arguments

:: HashMap Tp [Expr]

-> Expr

-> [[Tp]]
for each type instantiation, for each param, the input type as string

-> Interpreter (HashMap [Tp] String)

given sample inputs by type and type instantiations for a function, get its in/out pairs (by type)

fillHoles :: Int -> HashMap String Expr -> Set String -> [(String, Expr)] -> Expr -> Interpreter [Expr]

generate potential programs filling any holes in a given expression using some building blocks

fillHole :: HashMap String Expr -> Set String -> [(String, Expr)] -> Expr -> Interpreter ([(Expr, Set String, Expr)], [(Expr, Set String, Expr)])

filter building blocks to those matching a hole in the (let-in) expression, and get the results Exprs

genFn :: Int -> [(String, Expr)] -> HashMap String Expr -> Interpreter Expr

just directly sample a generated function, and see what types end up coming out.

genFns :: Int -> [(String, Expr)] -> HashMap String Expr -> Interpreter [Expr]

just directly generate any functions in batch, and see what types end up coming out. | in this approach, further nodes can impose new constraints on the type variables introduced in earlier nodes.

instantiateTypes :: HashMap Int [Tp] -> Tp -> Interpreter [Tp]

generate any combinations of a polymorphic type filled using a list of concrete types

instantiateTypeVars :: HashMap Int [Tp] -> HashMap String (Int, [Tp]) -> Interpreter [HashMap String Tp]

instantiate type variables

matchesConstraints :: Int -> Tp -> [Tp] -> Interpreter Bool

check if a type matches the given type constraints. | runs a command like `(undefined :: (Num a, Eq a) => a -> ()) (undefined :: Bool)`

matchesType :: Tp -> Tp -> Interpreter Bool

check if type a matches type b. | runs a command like `(undefined :: b -> ()) (undefined :: a)`. | without forall `a =>` constraints type variables will always match.

typeRelation :: Tp -> Tp -> Interpreter Ordering

find how two types relate | deprecated, not in use

dedupeFunctions :: HashMap Expr Tp -> HashMap Expr (HashMap [Tp] String) -> [Expr]

deduplicate functions by identical types + io, keeping the shortest | deprecated, not in use