Synthesis.Hint

Description

utility functions related to the Haskell Interpreter hint

Documentation

runInterpreterMain :: Interpreter () -> IO ()

run an interpreter monad, printing errors, ignoring values | deprecated, replace with interpretUnsafe

interpretUnsafe :: Interpreter a -> IO a

test an interpreter monad, printing errors, returning values

interpretSafe :: Interpreter a -> IO (Either InterpreterError a)

run an interpreter monad with imports

say :: String -> Interpreter ()

print in the Interpreter monad

debug :: String -> Interpreter ()

log: debug | deprecated, not in use

info :: String -> Interpreter ()

log: info | deprecated, not in use

notice :: String -> Interpreter ()

log: notice | deprecated, not in use

warning :: String -> Interpreter ()

log: warning

err :: String -> Interpreter ()

deprecated, not in use

log: error

critical :: String -> Interpreter ()

log: critical | deprecated, not in use

alert :: String -> Interpreter ()

log: alert | deprecated, not in use

emergency :: String -> Interpreter ()

log: emergency | deprecated, not in use

errorString :: InterpreterError -> String

run-time Language.Haskell.Interpreter compilation error

showError :: GhcError -> String

show a GHC error

interpretIO :: Bool -> String -> Interpreter (Either String String)

interpret a stringified IO command, either performing an additional typecheck (slower), or just crashing on error for a bogus Either

fnIoPairs :: Bool -> Int -> Expr -> Expr -> Interpreter String

get input-output pairs for a function given the inputs (for one concrete input type instantiation). | function application is run trough try-evaluate so as to Either-wrap potential run-time errors for partial functions. | the reason this function needs to be run through the interpreter is I only have the functioninputs as ASTstrings, | meaning I also only know the types at run-time (which is when my programs are constructed).

exprType :: Expr -> Interpreter Tp

get the type of an expression