Synthesis.Orphanage

Description

defined type class instances

Orphan instances

Pretty Tool

Instance details

Methods

pretty :: Tool -> Doc ann

prettyList :: [Tool] -> Doc ann

Store Boxed

Instance details

Methods

size :: Size Boxed

poke :: Boxed -> Poke ()

peek :: Peek Boxed

Store Tool

Instance details

Methods

size :: Size Tool

poke :: Tool -> Poke ()

peek :: Peek Tool

Store SrcSpan

Instance details

Methods

size :: Size SrcSpan

poke :: SrcSpan -> Poke ()

peek :: Peek SrcSpan

Store SrcSpanInfo

Instance details

Methods

size :: Size SrcSpanInfo

poke :: SrcSpanInfo -> Poke ()

peek :: Peek SrcSpanInfo

Store Stuff

Instance details

Methods

size :: Size Stuff

poke :: Stuff -> Poke ()

peek :: Peek Stuff

Hashable (Type l)
ensure I can use types as keys in HashMaps

Instance details

Methods

hashWithSalt :: Int -> Type l -> Int

hash :: Type l -> Int

Hashable (Exp l)
ensure I can use expressions as keys in HashMaps

Instance details

Methods

hashWithSalt :: Int -> Exp l -> Int

hash :: Exp l -> Int

Pretty (ModuleName l)

Instance details

Methods

pretty :: ModuleName l -> Doc ann

prettyList :: [ModuleName l] -> Doc ann

Pretty (SpecialCon l)

Instance details

Methods

pretty :: SpecialCon l -> Doc ann

prettyList :: [SpecialCon l] -> Doc ann

Pretty (QName l)

Instance details

Methods

pretty :: QName l -> Doc ann

prettyList :: [QName l] -> Doc ann

Pretty (Name l)

Instance details

Methods

pretty :: Name l -> Doc ann

prettyList :: [Name l] -> Doc ann

Pretty (IPName l)

Instance details

Methods

pretty :: IPName l -> Doc ann

prettyList :: [IPName l] -> Doc ann

Pretty (QOp l)

Instance details

Methods

pretty :: QOp l -> Doc ann

prettyList :: [QOp l] -> Doc ann

Pretty (Op l)

Instance details

Methods

pretty :: Op l -> Doc ann

prettyList :: [Op l] -> Doc ann

Pretty (CName l)

Instance details

Methods

pretty :: CName l -> Doc ann

prettyList :: [CName l] -> Doc ann

Pretty (Module l)

Instance details

Methods

pretty :: Module l -> Doc ann

prettyList :: [Module l] -> Doc ann

Pretty (ModuleHead l)

Instance details

Methods

pretty :: ModuleHead l -> Doc ann

prettyList :: [ModuleHead l] -> Doc ann

Pretty (ExportSpecList l)

Instance details

Methods

pretty :: ExportSpecList l -> Doc ann

prettyList :: [ExportSpecList l] -> Doc ann

Pretty (ExportSpec l)

Instance details

Methods

pretty :: ExportSpec l -> Doc ann

prettyList :: [ExportSpec l] -> Doc ann

Pretty (Namespace l)

Instance details

Methods

pretty :: Namespace l -> Doc ann

prettyList :: [Namespace l] -> Doc ann

Pretty (ImportDecl l)

Instance details

Methods

pretty :: ImportDecl l -> Doc ann

prettyList :: [ImportDecl l] -> Doc ann

Pretty (ImportSpecList l)

Instance details

Methods

pretty :: ImportSpecList l -> Doc ann

prettyList :: [ImportSpecList l] -> Doc ann

Pretty (ImportSpec l)

Instance details

Methods

pretty :: ImportSpec l -> Doc ann

prettyList :: [ImportSpec l] -> Doc ann

Pretty (Assoc l)

Instance details

Methods

pretty :: Assoc l -> Doc ann

prettyList :: [Assoc l] -> Doc ann

Pretty (Decl l)

Instance details

Methods

pretty :: Decl l -> Doc ann

prettyList :: [Decl l] -> Doc ann

Pretty (TypeEqn l)

Instance details

Methods

pretty :: TypeEqn l -> Doc ann

prettyList :: [TypeEqn l] -> Doc ann

Pretty (Annotation l)

Instance details

Methods

pretty :: Annotation l -> Doc ann

prettyList :: [Annotation l] -> Doc ann

Pretty (BooleanFormula l)

Instance details

Methods

pretty :: BooleanFormula l -> Doc ann

prettyList :: [BooleanFormula l] -> Doc ann

Pretty (Role l)

Instance details

Methods

pretty :: Role l -> Doc ann

prettyList :: [Role l] -> Doc ann

Pretty (DataOrNew l)

Instance details

Methods

pretty :: DataOrNew l -> Doc ann

prettyList :: [DataOrNew l] -> Doc ann

Pretty (InjectivityInfo l)

Instance details

Methods

pretty :: InjectivityInfo l -> Doc ann

prettyList :: [InjectivityInfo l] -> Doc ann

Pretty (ResultSig l)

Instance details

Methods

pretty :: ResultSig l -> Doc ann

prettyList :: [ResultSig l] -> Doc ann

Pretty (DeclHead l)

Instance details

Methods

pretty :: DeclHead l -> Doc ann

prettyList :: [DeclHead l] -> Doc ann

Pretty (InstRule l)

Instance details

Methods

pretty :: InstRule l -> Doc ann

prettyList :: [InstRule l] -> Doc ann

Pretty (InstHead l)

Instance details

Methods

pretty :: InstHead l -> Doc ann

prettyList :: [InstHead l] -> Doc ann

Pretty (Deriving l)

Instance details

Methods

pretty :: Deriving l -> Doc ann

prettyList :: [Deriving l] -> Doc ann

Pretty (DerivStrategy l)

Instance details

Methods

pretty :: DerivStrategy l -> Doc ann

prettyList :: [DerivStrategy l] -> Doc ann

Pretty (IPBind l)

Instance details

Methods

pretty :: IPBind l -> Doc ann

prettyList :: [IPBind l] -> Doc ann

Pretty (Match l)

Instance details

Methods

pretty :: Match l -> Doc ann

prettyList :: [Match l] -> Doc ann

Pretty (QualConDecl l)

Instance details

Methods

pretty :: QualConDecl l -> Doc ann

prettyList :: [QualConDecl l] -> Doc ann

Pretty (ConDecl l)

Instance details

Methods

pretty :: ConDecl l -> Doc ann

prettyList :: [ConDecl l] -> Doc ann

Pretty (FieldDecl l)

Instance details

Methods

pretty :: FieldDecl l -> Doc ann

prettyList :: [FieldDecl l] -> Doc ann

Pretty (GadtDecl l)

Instance details

Methods

pretty :: GadtDecl l -> Doc ann

prettyList :: [GadtDecl l] -> Doc ann

Pretty (ClassDecl l)

Instance details

Methods

pretty :: ClassDecl l -> Doc ann

prettyList :: [ClassDecl l] -> Doc ann

Pretty (InstDecl l)

Instance details

Methods

pretty :: InstDecl l -> Doc ann

prettyList :: [InstDecl l] -> Doc ann

Pretty (BangType l)

Instance details

Methods

pretty :: BangType l -> Doc ann

prettyList :: [BangType l] -> Doc ann

Pretty (Unpackedness l)

Instance details

Methods

pretty :: Unpackedness l -> Doc ann

prettyList :: [Unpackedness l] -> Doc ann

Pretty (Rhs l)

Instance details

Methods

pretty :: Rhs l -> Doc ann

prettyList :: [Rhs l] -> Doc ann

Pretty (GuardedRhs l)

Instance details

Methods

pretty :: GuardedRhs l -> Doc ann

prettyList :: [GuardedRhs l] -> Doc ann

Pretty (Type l)

Instance details

Methods

pretty :: Type l -> Doc ann

prettyList :: [Type l] -> Doc ann

Pretty (MaybePromotedName l)

Instance details

Methods

pretty :: MaybePromotedName l -> Doc ann

prettyList :: [MaybePromotedName l] -> Doc ann

Pretty (Promoted l)

Instance details

Methods

pretty :: Promoted l -> Doc ann

prettyList :: [Promoted l] -> Doc ann

Pretty (TyVarBind l)

Instance details

Methods

pretty :: TyVarBind l -> Doc ann

prettyList :: [TyVarBind l] -> Doc ann

Pretty (FunDep l)

Instance details

Methods

pretty :: FunDep l -> Doc ann

prettyList :: [FunDep l] -> Doc ann

Pretty (Context l)

Instance details

Methods

pretty :: Context l -> Doc ann

prettyList :: [Context l] -> Doc ann

Pretty (Asst l)

Instance details

Methods

pretty :: Asst l -> Doc ann

prettyList :: [Asst l] -> Doc ann

Pretty (Literal l)

Instance details

Methods

pretty :: Literal l -> Doc ann

prettyList :: [Literal l] -> Doc ann

Pretty (Exp l)

Instance details

Methods

pretty :: Exp l -> Doc ann

prettyList :: [Exp l] -> Doc ann

Pretty (XName l)

Instance details

Methods

pretty :: XName l -> Doc ann

prettyList :: [XName l] -> Doc ann

Pretty (XAttr l)

Instance details

Methods

pretty :: XAttr l -> Doc ann

prettyList :: [XAttr l] -> Doc ann

Pretty (Bracket l)

Instance details

Methods

pretty :: Bracket l -> Doc ann

prettyList :: [Bracket l] -> Doc ann

Pretty (Splice l)

Instance details

Methods

pretty :: Splice l -> Doc ann

prettyList :: [Splice l] -> Doc ann

Pretty (Safety l)

Instance details

Methods

pretty :: Safety l -> Doc ann

prettyList :: [Safety l] -> Doc ann

Pretty (CallConv l)

Instance details

Methods

pretty :: CallConv l -> Doc ann

prettyList :: [CallConv l] -> Doc ann

Pretty (ModulePragma l)

Instance details

Methods

pretty :: ModulePragma l -> Doc ann

prettyList :: [ModulePragma l] -> Doc ann

Pretty (Overlap l)

Instance details

Methods

pretty :: Overlap l -> Doc ann

prettyList :: [Overlap l] -> Doc ann

Pretty (Activation l)

Instance details

Methods

pretty :: Activation l -> Doc ann

prettyList :: [Activation l] -> Doc ann

Pretty (Rule l)

Instance details

Methods

pretty :: Rule l -> Doc ann

prettyList :: [Rule l] -> Doc ann

Pretty (RuleVar l)

Instance details

Methods

pretty :: RuleVar l -> Doc ann

prettyList :: [RuleVar l] -> Doc ann

Pretty (Pat l)

Instance details

Methods

pretty :: Pat l -> Doc ann

prettyList :: [Pat l] -> Doc ann

Pretty (PXAttr l)

Instance details

Methods

pretty :: PXAttr l -> Doc ann

prettyList :: [PXAttr l] -> Doc ann

Pretty (RPatOp l)

Instance details

Methods

pretty :: RPatOp l -> Doc ann

prettyList :: [RPatOp l] -> Doc ann

Pretty (RPat l)

Instance details

Methods

pretty :: RPat l -> Doc ann

prettyList :: [RPat l] -> Doc ann

Pretty (PatField l)

Instance details

Methods

pretty :: PatField l -> Doc ann

prettyList :: [PatField l] -> Doc ann

Pretty (Stmt l)

Instance details

Methods

pretty :: Stmt l -> Doc ann

prettyList :: [Stmt l] -> Doc ann

Pretty (QualStmt l)

Instance details

Methods

pretty :: QualStmt l -> Doc ann

prettyList :: [QualStmt l] -> Doc ann

Pretty (FieldUpdate l)

Instance details

Methods

pretty :: FieldUpdate l -> Doc ann

prettyList :: [FieldUpdate l] -> Doc ann

Pretty (Alt l)

Instance details

Methods

pretty :: Alt l -> Doc ann

prettyList :: [Alt l] -> Doc ann

Store l => Store (ModuleName l)

Instance details

Methods

size :: Size (ModuleName l)

poke :: ModuleName l -> Poke ()

peek :: Peek (ModuleName l)

Store l => Store (SpecialCon l)

Instance details

Methods

size :: Size (SpecialCon l)

poke :: SpecialCon l -> Poke ()

peek :: Peek (SpecialCon l)

Store l => Store (QName l)

Instance details

Methods

size :: Size (QName l)

poke :: QName l -> Poke ()

peek :: Peek (QName l)

Store l => Store (Name l)

Instance details

Methods

size :: Size (Name l)

poke :: Name l -> Poke ()

peek :: Peek (Name l)

Store l => Store (IPName l)

Instance details

Methods

size :: Size (IPName l)

poke :: IPName l -> Poke ()

peek :: Peek (IPName l)

Store l => Store (QOp l)

Instance details

Methods

size :: Size (QOp l)

poke :: QOp l -> Poke ()

peek :: Peek (QOp l)

Store l => Store (Op l)

Instance details

Methods

size :: Size (Op l)

poke :: Op l -> Poke ()

peek :: Peek (Op l)

Store l => Store (CName l)

Instance details

Methods

size :: Size (CName l)

poke :: CName l -> Poke ()

peek :: Peek (CName l)

Store l => Store (Module l)

Instance details

Methods

size :: Size (Module l)

poke :: Module l -> Poke ()

peek :: Peek (Module l)

Store l => Store (ModuleHead l)

Instance details

Methods

size :: Size (ModuleHead l)

poke :: ModuleHead l -> Poke ()

peek :: Peek (ModuleHead l)

Store l => Store (ExportSpecList l)

Instance details

Methods

size :: Size (ExportSpecList l)

poke :: ExportSpecList l -> Poke ()

peek :: Peek (ExportSpecList l)

Store l => Store (ExportSpec l)

Instance details

Methods

size :: Size (ExportSpec l)

poke :: ExportSpec l -> Poke ()

peek :: Peek (ExportSpec l)

Store l => Store (EWildcard l)

Instance details

Methods

size :: Size (EWildcard l)

poke :: EWildcard l -> Poke ()

peek :: Peek (EWildcard l)

Store l => Store (Namespace l)

Instance details

Methods

size :: Size (Namespace l)

poke :: Namespace l -> Poke ()

peek :: Peek (Namespace l)

Store l => Store (ImportDecl l)

Instance details

Methods

size :: Size (ImportDecl l)

poke :: ImportDecl l -> Poke ()

peek :: Peek (ImportDecl l)

Store l => Store (ImportSpecList l)

Instance details

Methods

size :: Size (ImportSpecList l)

poke :: ImportSpecList l -> Poke ()

peek :: Peek (ImportSpecList l)

Store l => Store (ImportSpec l)

Instance details

Methods

size :: Size (ImportSpec l)

poke :: ImportSpec l -> Poke ()

peek :: Peek (ImportSpec l)

Store l => Store (Assoc l)

Instance details

Methods

size :: Size (Assoc l)

poke :: Assoc l -> Poke ()

peek :: Peek (Assoc l)

Store l => Store (Decl l)

Instance details

Methods

size :: Size (Decl l)

poke :: Decl l -> Poke ()

peek :: Peek (Decl l)

Store l => Store (PatternSynDirection l)

Instance details

Methods

size :: Size (PatternSynDirection l)

poke :: PatternSynDirection l -> Poke ()

peek :: Peek (PatternSynDirection l)

Store l => Store (TypeEqn l)

Instance details

Methods

size :: Size (TypeEqn l)

poke :: TypeEqn l -> Poke ()

peek :: Peek (TypeEqn l)

Store l => Store (Annotation l)

Instance details

Methods

size :: Size (Annotation l)

poke :: Annotation l -> Poke ()

peek :: Peek (Annotation l)

Store l => Store (BooleanFormula l)

Instance details

Methods

size :: Size (BooleanFormula l)

poke :: BooleanFormula l -> Poke ()

peek :: Peek (BooleanFormula l)

Store l => Store (Role l)

Instance details

Methods

size :: Size (Role l)

poke :: Role l -> Poke ()

peek :: Peek (Role l)

Store l => Store (DataOrNew l)

Instance details

Methods

size :: Size (DataOrNew l)

poke :: DataOrNew l -> Poke ()

peek :: Peek (DataOrNew l)

Store l => Store (InjectivityInfo l)

Instance details

Methods

size :: Size (InjectivityInfo l)

poke :: InjectivityInfo l -> Poke ()

peek :: Peek (InjectivityInfo l)

Store l => Store (ResultSig l)

Instance details

Methods

size :: Size (ResultSig l)

poke :: ResultSig l -> Poke ()

peek :: Peek (ResultSig l)

Store l => Store (DeclHead l)

Instance details

Methods

size :: Size (DeclHead l)

poke :: DeclHead l -> Poke ()

peek :: Peek (DeclHead l)

Store l => Store (InstRule l)

Instance details

Methods

size :: Size (InstRule l)

poke :: InstRule l -> Poke ()

peek :: Peek (InstRule l)

Store l => Store (InstHead l)

Instance details

Methods

size :: Size (InstHead l)

poke :: InstHead l -> Poke ()

peek :: Peek (InstHead l)

Store l => Store (Deriving l)

Instance details

Methods

size :: Size (Deriving l)

poke :: Deriving l -> Poke ()

peek :: Peek (Deriving l)

Store l => Store (DerivStrategy l)

Instance details

Methods

size :: Size (DerivStrategy l)

poke :: DerivStrategy l -> Poke ()

peek :: Peek (DerivStrategy l)

Store l => Store (Binds l)

Instance details

Methods

size :: Size (Binds l)

poke :: Binds l -> Poke ()

peek :: Peek (Binds l)

Store l => Store (IPBind l)

Instance details

Methods

size :: Size (IPBind l)

poke :: IPBind l -> Poke ()

peek :: Peek (IPBind l)

Store l => Store (Match l)

Instance details

Methods

size :: Size (Match l)

poke :: Match l -> Poke ()

peek :: Peek (Match l)

Store l => Store (QualConDecl l)

Instance details

Methods

size :: Size (QualConDecl l)

poke :: QualConDecl l -> Poke ()

peek :: Peek (QualConDecl l)

Store l => Store (ConDecl l)

Instance details

Methods

size :: Size (ConDecl l)

poke :: ConDecl l -> Poke ()

peek :: Peek (ConDecl l)

Store l => Store (FieldDecl l)

Instance details

Methods

size :: Size (FieldDecl l)

poke :: FieldDecl l -> Poke ()

peek :: Peek (FieldDecl l)

Store l => Store (GadtDecl l)

Instance details

Methods

size :: Size (GadtDecl l)

poke :: GadtDecl l -> Poke ()

peek :: Peek (GadtDecl l)

Store l => Store (ClassDecl l)

Instance details

Methods

size :: Size (ClassDecl l)

poke :: ClassDecl l -> Poke ()

peek :: Peek (ClassDecl l)

Store l => Store (InstDecl l)

Instance details

Methods

size :: Size (InstDecl l)

poke :: InstDecl l -> Poke ()

peek :: Peek (InstDecl l)

Store l => Store (BangType l)

Instance details

Methods

size :: Size (BangType l)

poke :: BangType l -> Poke ()

peek :: Peek (BangType l)

Store l => Store (Unpackedness l)

Instance details

Methods

size :: Size (Unpackedness l)

poke :: Unpackedness l -> Poke ()

peek :: Peek (Unpackedness l)

Store l => Store (Rhs l)

Instance details

Methods

size :: Size (Rhs l)

poke :: Rhs l -> Poke ()

peek :: Peek (Rhs l)

Store l => Store (GuardedRhs l)

Instance details

Methods

size :: Size (GuardedRhs l)

poke :: GuardedRhs l -> Poke ()

peek :: Peek (GuardedRhs l)

Store l => Store (Type l)

Instance details

Methods

size :: Size (Type l)

poke :: Type l -> Poke ()

peek :: Peek (Type l)

Store l => Store (MaybePromotedName l)

Instance details

Methods

size :: Size (MaybePromotedName l)

poke :: MaybePromotedName l -> Poke ()

peek :: Peek (MaybePromotedName l)

Store l => Store (Promoted l)

Instance details

Methods

size :: Size (Promoted l)

poke :: Promoted l -> Poke ()

peek :: Peek (Promoted l)

Store l => Store (TyVarBind l)

Instance details

Methods

size :: Size (TyVarBind l)

poke :: TyVarBind l -> Poke ()

peek :: Peek (TyVarBind l)

Store l => Store (FunDep l)

Instance details

Methods

size :: Size (FunDep l)

poke :: FunDep l -> Poke ()

peek :: Peek (FunDep l)

Store l => Store (Context l)

Instance details

Methods

size :: Size (Context l)

poke :: Context l -> Poke ()

peek :: Peek (Context l)

Store l => Store (Asst l)

Instance details

Methods

size :: Size (Asst l)

poke :: Asst l -> Poke ()

peek :: Peek (Asst l)

Store l => Store (Literal l)

Instance details

Methods

size :: Size (Literal l)

poke :: Literal l -> Poke ()

peek :: Peek (Literal l)

Store l => Store (Sign l)

Instance details

Methods

size :: Size (Sign l)

poke :: Sign l -> Poke ()

peek :: Peek (Sign l)

Store l => Store (Exp l)

Instance details

Methods

size :: Size (Exp l)

poke :: Exp l -> Poke ()

peek :: Peek (Exp l)

Store l => Store (XName l)

Instance details

Methods

size :: Size (XName l)

poke :: XName l -> Poke ()

peek :: Peek (XName l)

Store l => Store (XAttr l)

Instance details

Methods

size :: Size (XAttr l)

poke :: XAttr l -> Poke ()

peek :: Peek (XAttr l)

Store l => Store (Bracket l)

Instance details

Methods

size :: Size (Bracket l)

poke :: Bracket l -> Poke ()

peek :: Peek (Bracket l)

Store l => Store (Splice l)

Instance details

Methods

size :: Size (Splice l)

poke :: Splice l -> Poke ()

peek :: Peek (Splice l)

Store l => Store (Safety l)

Instance details

Methods

size :: Size (Safety l)

poke :: Safety l -> Poke ()

peek :: Peek (Safety l)

Store l => Store (CallConv l)

Instance details

Methods

size :: Size (CallConv l)

poke :: CallConv l -> Poke ()

peek :: Peek (CallConv l)

Store l => Store (ModulePragma l)

Instance details

Methods

size :: Size (ModulePragma l)

poke :: ModulePragma l -> Poke ()

peek :: Peek (ModulePragma l)

Store l => Store (Overlap l)

Instance details

Methods

size :: Size (Overlap l)

poke :: Overlap l -> Poke ()

peek :: Peek (Overlap l)

Store l => Store (Activation l)

Instance details

Methods

size :: Size (Activation l)

poke :: Activation l -> Poke ()

peek :: Peek (Activation l)

Store l => Store (Rule l)

Instance details

Methods

size :: Size (Rule l)

poke :: Rule l -> Poke ()

peek :: Peek (Rule l)

Store l => Store (RuleVar l)

Instance details

Methods

size :: Size (RuleVar l)

poke :: RuleVar l -> Poke ()

peek :: Peek (RuleVar l)

Store l => Store (WarningText l)

Instance details

Methods

size :: Size (WarningText l)

poke :: WarningText l -> Poke ()

peek :: Peek (WarningText l)

Store l => Store (Pat l)

Instance details

Methods

size :: Size (Pat l)

poke :: Pat l -> Poke ()

peek :: Peek (Pat l)

Store l => Store (PXAttr l)

Instance details

Methods

size :: Size (PXAttr l)

poke :: PXAttr l -> Poke ()

peek :: Peek (PXAttr l)

Store l => Store (RPatOp l)

Instance details

Methods

size :: Size (RPatOp l)

poke :: RPatOp l -> Poke ()

peek :: Peek (RPatOp l)

Store l => Store (RPat l)

Instance details

Methods

size :: Size (RPat l)

poke :: RPat l -> Poke ()

peek :: Peek (RPat l)

Store l => Store (PatField l)

Instance details

Methods

size :: Size (PatField l)

poke :: PatField l -> Poke ()

peek :: Peek (PatField l)

Store l => Store (Stmt l)

Instance details

Methods

size :: Size (Stmt l)

poke :: Stmt l -> Poke ()

peek :: Peek (Stmt l)

Store l => Store (QualStmt l)

Instance details

Methods

size :: Size (QualStmt l)

poke :: QualStmt l -> Poke ()

peek :: Peek (QualStmt l)

Store l => Store (FieldUpdate l)

Instance details

Methods

size :: Size (FieldUpdate l)

poke :: FieldUpdate l -> Poke ()

peek :: Peek (FieldUpdate l)

Store l => Store (Alt l)

Instance details

Methods

size :: Size (Alt l)

poke :: Alt l -> Poke ()

peek :: Peek (Alt l)

(Pretty k, Pretty v) => Pretty (HashMap k v)
ensure I can print hashmaps

Instance details

Methods

pretty :: HashMap k v -> Doc ann

prettyList :: [HashMap k v] -> Doc ann