Synthesis.TypeGen

Description

functions related to type generation

Documentation

randomType :: Bool -> Bool -> Int -> HashMap String [Tp] -> Int -> IO Tp

randomly generate a type

randomFnType :: Bool -> Bool -> Int -> HashMap String [Tp] -> Int -> IO Tp

randomly generate a function type | deprecated, not in actual use (randomType is only used with allowFns as True in another deprecated function) TODO: ensure each type var is used at least twice

tyCon :: String -> Tp

create a monomorphic type node

tyApp :: Tp -> Tp -> Tp

create a polymorphic type node

genTypes :: Int -> Int -> IO (HashMap Int [Tp])

generate a number of concrete types to be used in type variable substitution

findTypeVars :: Tp -> HashMap String (Int, [Tp])

find the type variables and their constraints

fillTypeVars :: Tp -> HashMap String Tp -> Tp

substitute all type variable occurrences

mergeTyVars :: HashMap String [Tp] -> HashMap String [Tp] -> HashMap String [Tp]

deprecated, not in actual use (randomType is only used with allowFns as True in another deprecated function)

unParseResult :: ParseResult a -> Either String a

typeSane :: Tp -> Bool

check if a type is sane -- basically we wanna throw out crap like list of function

isFn :: Tp -> Bool

check if a type is a function type

hasFn :: Tp -> Bool

check if a type contains a function type

nubPp :: Pretty a => [a] -> [a]

filter out duplicate types. note this dedupe will fail for type variable variations...

unList :: Expr -> [Expr]

unpack a list expression