Synthesis.Types

Description

utility functions specifically related to types

Documentation

type Tp = Type L

type Expr = Exp L

type Hole = SpecialCon L

randomType :: Bool -> Bool -> Int -> HashMap String [Tp] -> Int -> IO Tp

randomly generate a type TODO: allow generating new type vars

randomFnType :: Bool -> Bool -> Int -> HashMap String [Tp] -> Int -> IO Tp

randomly generate a function type TODO: ensure each type var is used at least twice

tyCon :: String -> Tp

create a monomorphic type node

tyApp :: Tp -> Tp -> Tp

create a polymorphic type node

fnTypeIO :: Tp -> ([Tp], Tp)

extract the input and output types from a function type | deprecated, not in use TODO: Maybe

genTypes :: Int -> Int -> IO (Item Tp)

generate a number of concrete types to be used in type variable substitution TODO: move the flatten/nub in

holeType :: Expr -> Tp

this function takes an explicitly typed hole, returning its type | deprecated, not in use TODO: Maybe

var :: String -> Expr

create a variable node

tyVar :: String -> Tp

create a type variable node

qName :: String -> QName L

create a qname node

l :: L

dummy source span info, because I don't care

findTypeVars :: Tp -> HashMap String [Tp]

find the type variables and their occurrences

fillTypeVars :: Tp -> HashMap String Tp -> Tp

substitute all type variable occurrences

star :: Tp

star type node: * | deprecated, not in use

wildcard :: Tp

wildcard type node: _

expTypeSig :: Expr -> Tp -> Expr

annotate an expression node with a type signature

tyFun :: Tp -> Tp -> Tp

type for a function

letIn :: HashMap String Expr -> Expr -> Expr

let-expression

app :: Expr -> Expr -> Expr

function application

parseExpr :: String -> Expr

parse an expression from a string

parseType :: String -> Tp

parse a type from a string

undef :: Tp -> Expr

create a typed expression without value, intended for checking types

cxTuple :: [Asst L] -> Context L

tuple of type constraints

classA :: QName L -> [Tp] -> Asst L

type constraint assertion

tyForall :: Maybe [TyVarBind L] -> Maybe (Context L) -> Tp -> Tp

type constraint

mergeTyVars :: HashMap String [Tp] -> HashMap String [Tp] -> HashMap String [Tp]

unParseResult :: ParseResult a -> Either String a

unit :: Tp

unit type: ()

symbol :: String -> QOp L

symbol for use in infix expressions

pvar :: String -> Pat L

variable name as used on the left-hand side of an assignment

ptuple :: [Pat L] -> Pat L

tuple pattern as used on the left-hand side of an assignment

paren :: Expr -> Expr

parenthesized expression

infixApp :: Expr -> QOp L -> Expr -> Expr

infix function application

dollar :: QOp L

$

dot :: QOp L

.

list :: [Expr] -> Expr

a list of expressions

tuple :: [Expr] -> Expr

a tuple of expressions

int :: Integer -> Expr

Int expression

string :: String -> Expr

String expression

con :: String -> Expr

data constructor

lambda :: [Pat L] -> Expr -> Expr

lambda function

tyList :: Tp -> Tp

list type

fnInputTypes :: Tp -> [Tp]

extract the input types from a function type

isFn :: Tp -> Bool

check if a type is a function type

hasFn :: Tp -> Bool

check if a type contains a function type

nubPp :: Pretty a => [a] -> [a]

filter out duplicate types. note this dedupe will fail for type variable variations...