Synthesis.Utility

Description

utility functions

Documentation

data Item a

a homogeneous nested list

Constructors

One [a]

Many [Item a]

data NestedTuple a

a homogeneous nested tuple

Constructors

SingleTuple (a, a)

DeepTuple (a, NestedTuple a)

flatten :: Item a -> [a]

flatten a nested list

pick :: [a] -> IO a

randomly pick an item from a list

mapKeys :: (Hashable k, Eq k, Hashable k_, Eq k_) => (k -> k_) -> HashMap k v -> HashMap k_ v

map over the keys of a hashmap | deprecated, not in use

groupByVal :: (Hashable v, Ord v) => [(k, v)] -> HashMap v [k]

group a list of k/v pairs by values, essentially inverting the original HashMap

fromKeys :: (Hashable k, Eq k) => (k -> v) -> [k] -> HashMap k v

create a HashMap by mapping over a list of keys

fromVals :: (Hashable k, Eq k) => (v -> k) -> [v] -> HashMap k v

create a HashMap by mapping over a list of values | deprecated, not in use

flattenTuple :: NestedTuple a -> [a]

flatten a nested tuple | deprecated, not in use

mapTuple :: (a -> b) -> (a, a) -> (b, b)

map over both elements of a tuple | deprecated, not in use

mapTuple3 :: (a -> b) -> (a, a, a) -> (b, b, b)

map over of a 3-element tuple

tuplify3 :: [a] -> (a, a, a)

convert a list of 3(+) items to a tuple of 3

untuple3 :: (a, a, a) -> [a]

unpack a tuple into a list

while :: Monad m => (a -> Bool) -> (a -> m a) -> a -> m a

while a predicate holds, perform a monadic operation starting from an initial value | deprecated, not in use

pp :: Pretty a => a -> String

shorthand for pretty-printing AST nodes, used for comparisons

pp_ :: Pretty a => a -> String

shorthand for pretty-printing AST nodes, used in my top-level module

pickKeys :: (Hashable k, Eq k) => [k] -> HashMap k v -> HashMap k v

pick some keys from a hashmap

composeSetters :: (s -> s -> s_) -> (s -> t) -> (t -> b -> s) -> s -> b -> s_

compose two setter functions, as I didn't figure out lenses and their monad instantiations

fisherYates :: RandomGen g => g -> [a] -> ([a], g)

randomly shuffle a list

randomSplit :: RandomGen g => g -> (Double, Double, Double) -> [a] -> ([a], [a], [a])

randomly split a dataset into subsets based on the indicated split ratio

flipOrder :: Ordering -> Ordering

flip an Ordering | deprecated, not in use

equating :: Eq a => (b -> a) -> b -> b -> Bool

mapped equality check

fromKeysM :: (Monad m, Hashable k, Eq k) => (k -> m v) -> [k] -> m (HashMap k v)

create a monadic HashMap by mapping over a list of keys

fromValsM :: (Monad m, Hashable k, Eq k) => (v -> m k) -> [v] -> m (HashMap k v)

create a monadic HashMap by mapping over a list of values

ppMap :: (Pretty k, Show v) => HashMap k v -> String

show drop-in for HashMap with Pretty keys. | deprecated, not in use

filterHmM :: (Monad m, Hashable k, Eq k) => ((k, v) -> m Bool) -> HashMap k v -> m (HashMap k v)

deprecated, not in use

pickKeysSafe :: (Hashable k, Eq k) => [k] -> HashMap k v -> HashMap k v

pick some keys from a hashmap

nest :: Monad m => Int -> (a -> m a) -> a -> m a

monadic version of nTimes