Synthesis.Orphanage

Description

defined type class instances

Orphan instances

Pretty Tool

Instance details

Methods

pretty :: Tool -> Doc ann

prettyList :: [Tool] -> Doc ann

Hashable (Type l)
ensure I can use types as keys in HashMaps

Instance details

Methods

hashWithSalt :: Int -> Type l -> Int

hash :: Type l -> Int

Hashable (Exp l)
ensure I can use expressions as keys in HashMaps

Instance details

Methods

hashWithSalt :: Int -> Exp l -> Int

hash :: Exp l -> Int

Pretty (ModuleName l)

Instance details

Methods

pretty :: ModuleName l -> Doc ann

prettyList :: [ModuleName l] -> Doc ann

Pretty (SpecialCon l)

Instance details

Methods

pretty :: SpecialCon l -> Doc ann

prettyList :: [SpecialCon l] -> Doc ann

Pretty (QName l)

Instance details

Methods

pretty :: QName l -> Doc ann

prettyList :: [QName l] -> Doc ann

Pretty (Name l)

Instance details

Methods

pretty :: Name l -> Doc ann

prettyList :: [Name l] -> Doc ann

Pretty (IPName l)

Instance details

Methods

pretty :: IPName l -> Doc ann

prettyList :: [IPName l] -> Doc ann

Pretty (QOp l)

Instance details

Methods

pretty :: QOp l -> Doc ann

prettyList :: [QOp l] -> Doc ann

Pretty (Op l)

Instance details

Methods

pretty :: Op l -> Doc ann

prettyList :: [Op l] -> Doc ann

Pretty (CName l)

Instance details

Methods

pretty :: CName l -> Doc ann

prettyList :: [CName l] -> Doc ann

Pretty (Module l)

Instance details

Methods

pretty :: Module l -> Doc ann

prettyList :: [Module l] -> Doc ann

Pretty (ModuleHead l)

Instance details

Methods

pretty :: ModuleHead l -> Doc ann

prettyList :: [ModuleHead l] -> Doc ann

Pretty (ExportSpecList l)

Instance details

Methods

pretty :: ExportSpecList l -> Doc ann

prettyList :: [ExportSpecList l] -> Doc ann

Pretty (ExportSpec l)

Instance details

Methods

pretty :: ExportSpec l -> Doc ann

prettyList :: [ExportSpec l] -> Doc ann

Pretty (Namespace l)

Instance details

Methods

pretty :: Namespace l -> Doc ann

prettyList :: [Namespace l] -> Doc ann

Pretty (ImportDecl l)

Instance details

Methods

pretty :: ImportDecl l -> Doc ann

prettyList :: [ImportDecl l] -> Doc ann

Pretty (ImportSpecList l)

Instance details

Methods

pretty :: ImportSpecList l -> Doc ann

prettyList :: [ImportSpecList l] -> Doc ann

Pretty (ImportSpec l)

Instance details

Methods

pretty :: ImportSpec l -> Doc ann

prettyList :: [ImportSpec l] -> Doc ann

Pretty (Assoc l)

Instance details

Methods

pretty :: Assoc l -> Doc ann

prettyList :: [Assoc l] -> Doc ann

Pretty (Decl l)

Instance details

Methods

pretty :: Decl l -> Doc ann

prettyList :: [Decl l] -> Doc ann

Pretty (TypeEqn l)

Instance details

Methods

pretty :: TypeEqn l -> Doc ann

prettyList :: [TypeEqn l] -> Doc ann

Pretty (Annotation l)

Instance details

Methods

pretty :: Annotation l -> Doc ann

prettyList :: [Annotation l] -> Doc ann

Pretty (BooleanFormula l)

Instance details

Methods

pretty :: BooleanFormula l -> Doc ann

prettyList :: [BooleanFormula l] -> Doc ann

Pretty (Role l)

Instance details

Methods

pretty :: Role l -> Doc ann

prettyList :: [Role l] -> Doc ann

Pretty (DataOrNew l)

Instance details

Methods

pretty :: DataOrNew l -> Doc ann

prettyList :: [DataOrNew l] -> Doc ann

Pretty (InjectivityInfo l)

Instance details

Methods

pretty :: InjectivityInfo l -> Doc ann

prettyList :: [InjectivityInfo l] -> Doc ann

Pretty (ResultSig l)

Instance details

Methods

pretty :: ResultSig l -> Doc ann

prettyList :: [ResultSig l] -> Doc ann

Pretty (DeclHead l)

Instance details

Methods

pretty :: DeclHead l -> Doc ann

prettyList :: [DeclHead l] -> Doc ann

Pretty (InstRule l)

Instance details

Methods

pretty :: InstRule l -> Doc ann

prettyList :: [InstRule l] -> Doc ann

Pretty (InstHead l)

Instance details

Methods

pretty :: InstHead l -> Doc ann

prettyList :: [InstHead l] -> Doc ann

Pretty (Deriving l)

Instance details

Methods

pretty :: Deriving l -> Doc ann

prettyList :: [Deriving l] -> Doc ann

Pretty (DerivStrategy l)

Instance details

Methods

pretty :: DerivStrategy l -> Doc ann

prettyList :: [DerivStrategy l] -> Doc ann

Pretty (IPBind l)

Instance details

Methods

pretty :: IPBind l -> Doc ann

prettyList :: [IPBind l] -> Doc ann

Pretty (Match l)

Instance details

Methods

pretty :: Match l -> Doc ann

prettyList :: [Match l] -> Doc ann

Pretty (QualConDecl l)

Instance details

Methods

pretty :: QualConDecl l -> Doc ann

prettyList :: [QualConDecl l] -> Doc ann

Pretty (ConDecl l)

Instance details

Methods

pretty :: ConDecl l -> Doc ann

prettyList :: [ConDecl l] -> Doc ann

Pretty (FieldDecl l)

Instance details

Methods

pretty :: FieldDecl l -> Doc ann

prettyList :: [FieldDecl l] -> Doc ann

Pretty (GadtDecl l)

Instance details

Methods

pretty :: GadtDecl l -> Doc ann

prettyList :: [GadtDecl l] -> Doc ann

Pretty (ClassDecl l)

Instance details

Methods

pretty :: ClassDecl l -> Doc ann

prettyList :: [ClassDecl l] -> Doc ann

Pretty (InstDecl l)

Instance details

Methods

pretty :: InstDecl l -> Doc ann

prettyList :: [InstDecl l] -> Doc ann

Pretty (BangType l)

Instance details

Methods

pretty :: BangType l -> Doc ann

prettyList :: [BangType l] -> Doc ann

Pretty (Unpackedness l)

Instance details

Methods

pretty :: Unpackedness l -> Doc ann

prettyList :: [Unpackedness l] -> Doc ann

Pretty (Rhs l)

Instance details

Methods

pretty :: Rhs l -> Doc ann

prettyList :: [Rhs l] -> Doc ann

Pretty (GuardedRhs l)

Instance details

Methods

pretty :: GuardedRhs l -> Doc ann

prettyList :: [GuardedRhs l] -> Doc ann

Pretty (Type l)

Instance details

Methods

pretty :: Type l -> Doc ann

prettyList :: [Type l] -> Doc ann

Pretty (MaybePromotedName l)

Instance details

Methods

pretty :: MaybePromotedName l -> Doc ann

prettyList :: [MaybePromotedName l] -> Doc ann

Pretty (Promoted l)

Instance details

Methods

pretty :: Promoted l -> Doc ann

prettyList :: [Promoted l] -> Doc ann

Pretty (TyVarBind l)

Instance details

Methods

pretty :: TyVarBind l -> Doc ann

prettyList :: [TyVarBind l] -> Doc ann

Pretty (FunDep l)

Instance details

Methods

pretty :: FunDep l -> Doc ann

prettyList :: [FunDep l] -> Doc ann

Pretty (Context l)

Instance details

Methods

pretty :: Context l -> Doc ann

prettyList :: [Context l] -> Doc ann

Pretty (Asst l)

Instance details

Methods

pretty :: Asst l -> Doc ann

prettyList :: [Asst l] -> Doc ann

Pretty (Literal l)

Instance details

Methods

pretty :: Literal l -> Doc ann

prettyList :: [Literal l] -> Doc ann

Pretty (Exp l)

Instance details

Methods

pretty :: Exp l -> Doc ann

prettyList :: [Exp l] -> Doc ann

Pretty (XName l)

Instance details

Methods

pretty :: XName l -> Doc ann

prettyList :: [XName l] -> Doc ann

Pretty (XAttr l)

Instance details

Methods

pretty :: XAttr l -> Doc ann

prettyList :: [XAttr l] -> Doc ann

Pretty (Bracket l)

Instance details

Methods

pretty :: Bracket l -> Doc ann

prettyList :: [Bracket l] -> Doc ann

Pretty (Splice l)

Instance details

Methods

pretty :: Splice l -> Doc ann

prettyList :: [Splice l] -> Doc ann

Pretty (Safety l)

Instance details

Methods

pretty :: Safety l -> Doc ann

prettyList :: [Safety l] -> Doc ann

Pretty (CallConv l)

Instance details

Methods

pretty :: CallConv l -> Doc ann

prettyList :: [CallConv l] -> Doc ann

Pretty (ModulePragma l)

Instance details

Methods

pretty :: ModulePragma l -> Doc ann

prettyList :: [ModulePragma l] -> Doc ann

Pretty (Overlap l)

Instance details

Methods

pretty :: Overlap l -> Doc ann

prettyList :: [Overlap l] -> Doc ann

Pretty (Activation l)

Instance details

Methods

pretty :: Activation l -> Doc ann

prettyList :: [Activation l] -> Doc ann

Pretty (Rule l)

Instance details

Methods

pretty :: Rule l -> Doc ann

prettyList :: [Rule l] -> Doc ann

Pretty (RuleVar l)

Instance details

Methods

pretty :: RuleVar l -> Doc ann

prettyList :: [RuleVar l] -> Doc ann

Pretty (Pat l)

Instance details

Methods

pretty :: Pat l -> Doc ann

prettyList :: [Pat l] -> Doc ann

Pretty (PXAttr l)

Instance details

Methods

pretty :: PXAttr l -> Doc ann

prettyList :: [PXAttr l] -> Doc ann

Pretty (RPatOp l)

Instance details

Methods

pretty :: RPatOp l -> Doc ann

prettyList :: [RPatOp l] -> Doc ann

Pretty (RPat l)

Instance details

Methods

pretty :: RPat l -> Doc ann

prettyList :: [RPat l] -> Doc ann

Pretty (PatField l)

Instance details

Methods

pretty :: PatField l -> Doc ann

prettyList :: [PatField l] -> Doc ann

Pretty (Stmt l)

Instance details

Methods

pretty :: Stmt l -> Doc ann

prettyList :: [Stmt l] -> Doc ann

Pretty (QualStmt l)

Instance details

Methods

pretty :: QualStmt l -> Doc ann

prettyList :: [QualStmt l] -> Doc ann

Pretty (FieldUpdate l)

Instance details

Methods

pretty :: FieldUpdate l -> Doc ann

prettyList :: [FieldUpdate l] -> Doc ann

Pretty (Alt l)

Instance details

Methods

pretty :: Alt l -> Doc ann

prettyList :: [Alt l] -> Doc ann

(Pretty k, Pretty v) => Pretty (HashMap k v)
ensure I can print lists

Instance details

Methods

pretty :: HashMap k v -> Doc ann

prettyList :: [HashMap k v] -> Doc ann